# tests/snapshots/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "apex-admin",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/auth/login": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "账号密码登录",
        "description": "账号密码登录 — SPEC 12.1 / 12.2 / 12.4.\n\nSPEC 12.1: 登录成功创建服务端会话，返回不透明 Access Token。\nSPEC 12.2: 登录成功创建 Refresh Token Family，经 Set-Cookie 下发。\nSPEC 12.4: 登录响应必须设置 ``Cache-Control: no-store``。\nSPEC 12.1: Access Token 仅在响应体中返回一次。\nSPEC 12.4: Refresh Token 仅经 Set-Cookie 下发，不进入 JSON 响应。\n\n登录失败返回 401（``AUTH.INVALID_CREDENTIALS``），所有失败路径\n返回完全一致的响应，防止账号枚举（SPEC 12.4）。",
        "operationId": "auth_login",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": []
      }
    },
    "/api/v1/auth/refresh": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "刷新 Access Token",
        "description": "Refresh Token 轮换 — SPEC 12.2 / 12.4.\n\nSPEC 12.2: 使用 Refresh Token 获取新的 Access Token。\n旧 Refresh Token 立即失效，新 Refresh Token 仅经 Set-Cookie 下发。\nSPEC 12.4: 校验 Origin 精确匹配白名单（防止 CSRF）。\nSPEC 12.4: 刷新响应必须设置 ``Cache-Control: no-store``。\nSPEC 12.2: Refresh Token 不进入 JSON 响应。\n\n刷新失败返回 401（``AUTH.REFRESH_FAILED``），不泄露失败原因。",
        "operationId": "auth_refresh",
        "parameters": [
          {
            "name": "__Host-apex_refresh",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "  Host-Apex Refresh"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RefreshResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": []
      }
    },
...
    "/api/v1/auth/users/{userId}/force-offline": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "管理员强制用户下线",
        "description": "管理员强制用户下线 — SPEC 12.3 / 18.1.\n\nSPEC 12.3: \"管理员可以强制用户下线\"。\n吊销目标用户的全部活动会话，使目标用户下一个请求立即收到 401。\n记录登录日志（SPEC 18.1）。\n\n此端点属于管理接口，通过权限依赖保护（SPEC 23.5）。",
        "operationId": "auth_force_offline",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "目标用户 ID",
              "title": "Userid"
            },
            "description": "目标用户 ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ForceOfflineResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ForceOfflineResponse": {
        "properties": {
          "userId": {
            "type": "string",
            "title": "Userid",
            "description": "目标用户 ID"
          },
          "revokedSessions": {
            "type": "integer",
            "title": "Revokedsessions",
            "description": "被吊销的会话数量"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "userId",
          "revokedSessions"
        ],
        "title": "ForceOfflineResponse",
        "description": "管理员强制用户下线响应 — SPEC 12.3.\n\n属性:\n    user_id:           被强制下线的目标用户 ID。\n    revoked_sessions:  被吊销的活动会话数量。"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LoginRequest": {
        "properties": {
          "username": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Username",
            "description": "用户名/登录账号"
          },
          "password": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "Password",
            "description": "明文密码"
          },
          "device": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200
              },
              {
                "type": "null"
              }
            ],
            "title": "Device",
            "description": "设备标识（可选）"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "username",
          "password"
        ],
        "title": "LoginRequest",
        "description": "登录请求 — SPEC 12.1 / 9.2.\n\nSPEC 12.1: \"用户名和密码登录\"。\n继承 ``StrictBaseModel``（``extra=\"forbid\"``），携带未知字段返回 422。\n\n属性:\n    username: 用户名/登录账号。\n    password: 明文密码（不在日志中记录，SPEC 23.2 / 12.4）。\n    device:   设备标识（可选），用于会话记录设备信息。"
      },
      "LoginResponse": {
        "properties": {
          "accessToken": {
            "type": "string",
            "title": "Accesstoken",
            "description": "不透明 Access Token（仅返回一次）"
          },
          "tokenType": {
            "type": "string",
            "title": "Tokentype",
            "description": "Token 类型",
            "default": "Bearer"
          },
          "expiresIn": {
            "type": "integer",
            "title": "Expiresin",
            "description": "Token 有效期（秒）"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "accessToken",
          "expiresIn"
        ],
        "title": "LoginResponse",
        "description": "登录成功响应 — SPEC 12.1.\n\nSPEC 12.1: \"Access Token 仅在登录或刷新响应体中返回一次\"。\nSPEC 12.1: \"G2 固定使用不透明随机 Bearer Access Token，不使用 JWT\"。\n``access_token`` 为不透明随机字符串，仅在本次响应中返回，\n前端只允许保存在内存中。\n\nSPEC 12.4: 登录响应必须设置 ``Cache-Control: no-store``\n（由 Router 设置响应头，不在响应体中）。\n\n属性:\n    access_token: 不透明 Access Token 字符串（仅在响应体中返回一次）。\n    token_type:   固定为 ``\"Bearer\"``。\n    expires_in:   Token 有效期（秒），默认 900（15 分钟）。"
      },
...
      "RefreshResponse": {
        "properties": {
          "accessToken": {
            "type": "string",
            "title": "Accesstoken",
            "description": "新不透明 Access Token"
          },
          "tokenType": {
            "type": "string",
            "title": "Tokentype",
            "description": "Token 类型",
            "default": "Bearer"
          },
          "expiresIn": {
            "type": "integer",
            "title": "Expiresin",
            "description": "Token 有效期（秒）"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "accessToken",
          "expiresIn"
        ],
        "title": "RefreshResponse",
        "description": "刷新响应 — SPEC 12.2.\n\nSPEC 12.2: 刷新成功时响应体仅包含新 Access Token。新 Refresh Token\n仅经 Set-Cookie 下发，不进入 JSON 响应（SPEC 12.2 / 12.4）。\nSPEC 12.4: 刷新响应必须设置 ``Cache-Control: no-store``\n（由 Router 设置响应头）。\n\n属性:\n    access_token: 新不透明 Access Token 字符串。\n    token_type:   固定为 ``\"Bearer\"``。\n    expires_in:   Token 有效期（秒），默认 900（15 分钟）。"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "Opaque",
        "description": "粘贴登录接口返回的 access_token（无需手写 Bearer 前缀）。认证依赖从 Authorization: Bearer <token> 提取不透明 Access Token（SPEC 12.1）。"
      }
    }
  },
...
  "security": [
    {
      "BearerAuth": []
    }
  ]
}